updated tut 3

+++ tut03/tut03.py
@@ -3,23 +3,22 @@
 def output_by_subject() :
     import os
     #os.mkdir("output_by_subject")
-    try : 
-        os.mkdir("output_by_subject")
-    except  FileExistsError:
-        pass    
+    if(os.path.exists("output_by_subject")) : 
+        pass
+    else :
+            os.mkdir("output_by_subject")
     with open("regtable_old.csv", "r") as f:
         for line in f:
             file1 = line.split(',')
             del file1[4:8]
             del file1[2]
             if (file1[2] =="subno"):continue
-            try : 
-                with open(f"output_by_subject\\{file1[2]}.csv" ):
+            if(os.path.exists(f"output_by_subject\\{file1[2]}.csv")): 
                     with open(f"output_by_subject\\{file1[2]}.csv", "a") as f11:
                         file1=",".join(file1)
                         f11.write(file1)
 
-            except IOError:
+            else :
                 with open(f"output_by_subject\\{file1[2]}.csv", "w") as f11:
                         f11.write("rollno,register_sem,subno,sub_type\n")
                         file1=",".join(file1)
@@ -29,22 +28,21 @@ def output_by_subject() :
 
 def output_individual_roll():
     import os
-    try : 
-        os.mkdir("output_individual_roll")
-    except  FileExistsError:
+    if(os.path.exists("output_individual_roll")) : 
         pass
+    else  :
+        os.mkdir("output_individual_roll")
     with open("regtable_old.csv", "r") as f:
         for line in f:
             file1 = line.split(',')
             del file1[4:8]
             del file1[2]
             if (file1[0] =="rollno"):continue
-            try: 
-                with open(f"output_individual_roll\\{file1[0]}.csv"):
-                    with open(f"output_individual_roll\\{file1[0]}.csv", "a") as f11:
+            if(os.path.exists(f"output_individual_roll\\{file1[0]}.csv")) :
+                with open(f"output_individual_roll\\{file1[0]}.csv", "a") as f11:
                         file1=",".join(file1)
                         f11.write(file1)
-            except IOError:
+            else :
                 with open(f"output_individual_roll\\{file1[0]}.csv", "w") as f11:
                         f11.write("rollno,register_sem,subno,sub_type\n")
                         file1=",".join(file1)
